chore: Move cmd/searcher/protocol to internal/searcher/protocol
Camden and I noticed while hacking together today that we import the searcher protocol across cmd/ boundaries, so this PR moves it to internal, which is in line with what we do for gitserver.

Test plan:

Bazel and go compiler don't complain.
