+++ cmd/searcher/internal/search/BUILD.bazel
@@ -25,7 +25,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/cmd/searcher/internal/search",
     visibility = ["//cmd/searcher:__subpackages__"],
     deps = [
-        "//cmd/searcher/protocol",
         "//internal/api",
         "//internal/comby",
         "//internal/conf",
@@ -42,6 +41,7 @@ go_library(
         "//internal/search/casetransform",
         "//internal/search/query",
         "//internal/search/zoektquery",
+        "//internal/searcher/protocol",
         "//internal/searcher/v1:searcher",
         "//internal/trace",
         "//lib/codeintel/languages",
@@ -135,7 +135,6 @@ go_test(
     ],
     embed = [":search"],
     deps = [
-        "//cmd/searcher/protocol",
         "//internal/api",
         "//internal/comby",
         "//internal/conf",
@@ -146,6 +145,7 @@ go_test(
         "//internal/grpc/defaults",
         "//internal/observation",
         "//internal/search/backend",
+        "//internal/searcher/protocol",
         "//internal/searcher/v1:searcher",
         "//lib/errors",
         "//schema",
