+++ cmd/searcher/internal/search/pathmatch.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/grafana/regexp"
 
-	"github.com/sourcegraph/sourcegraph/cmd/searcher/protocol"
 	"github.com/sourcegraph/sourcegraph/internal/search/query"
+	"github.com/sourcegraph/sourcegraph/internal/searcher/protocol"
 )
 
 type pathMatcher struct {
