+++ cmd/searcher/internal/search/search_grpc.go
@@ -9,7 +9,7 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	"github.com/sourcegraph/sourcegraph/cmd/searcher/protocol"
+	"github.com/sourcegraph/sourcegraph/internal/searcher/protocol"
 	proto "github.com/sourcegraph/sourcegraph/internal/searcher/v1"
 )
