+++ cmd/searcher/internal/search/sender_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/sourcegraph/conc/pool"
 	"github.com/stretchr/testify/require"
 
-	"github.com/sourcegraph/sourcegraph/cmd/searcher/protocol"
+	"github.com/sourcegraph/sourcegraph/internal/searcher/protocol"
 )
 
 func TestLimitedStream(t *testing.T) {
