+++ internal/search/BUILD.bazel
@@ -15,7 +15,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/search",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//cmd/searcher/protocol",
         "//internal/api",
         "//internal/collections",
         "//internal/conf",
@@ -31,6 +30,7 @@ go_library(
         "//internal/search/query",
         "//internal/search/result",
         "//internal/search/streaming/http",
+        "//internal/searcher/protocol",
         "//internal/trace",
         "//internal/trace/policy",
         "//internal/types",
