+++ internal/search/job/jobutil/BUILD.bazel
@@ -22,7 +22,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/search/job/jobutil",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//cmd/searcher/protocol",
         "//internal/actor",
         "//internal/api",
         "//internal/authz",
@@ -51,6 +50,7 @@ go_library(
         "//internal/search/streaming",
         "//internal/search/structural",
         "//internal/search/zoekt",
+        "//internal/searcher/protocol",
         "//internal/telemetry",
         "//internal/telemetry/teestore",
         "//internal/telemetry/telemetryrecorder",
@@ -90,7 +90,6 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":jobutil"],
     deps = [
-        "//cmd/searcher/protocol",
         "//internal/actor",
         "//internal/api",
         "//internal/authz",
@@ -112,6 +111,7 @@ go_test(
         "//internal/search/searcher",
         "//internal/search/streaming",
         "//internal/search/zoekt",
+        "//internal/searcher/protocol",
         "//internal/telemetry/telemetrytest",
         "//internal/types",
         "//lib/errors",
