+++ internal/search/repos/BUILD.bazel
@@ -11,7 +11,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/search/repos",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//cmd/searcher/protocol",
         "//internal/api",
         "//internal/conf",
         "//internal/database",
@@ -28,6 +27,7 @@ go_library(
         "//internal/search/searcher",
         "//internal/search/streaming",
         "//internal/search/zoekt",
+        "//internal/searcher/protocol",
         "//internal/trace",
         "//internal/types",
         "//lib/errors",
@@ -57,7 +57,6 @@ go_test(
         "requires-network",
     ],
     deps = [
-        "//cmd/searcher/protocol",
         "//internal/api",
         "//internal/database",
         "//internal/database/dbmocks",
@@ -71,6 +70,7 @@ go_test(
         "//internal/search/query",
         "//internal/search/searcher",
         "//internal/search/streaming",
+        "//internal/searcher/protocol",
         "//internal/types",
         "//lib/errors",
         "//lib/iterator",
