+++ internal/search/searcher/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/search/searcher",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//cmd/searcher/protocol",
         "//internal/api",
         "//internal/conf",
         "//internal/endpoint",
@@ -24,6 +23,7 @@ go_library(
         "//internal/search/job",
         "//internal/search/result",
         "//internal/search/streaming",
+        "//internal/searcher/protocol",
         "//internal/searcher/v1:searcher",
         "//internal/symbols",
         "//internal/trace",
