+++ internal/searcher/protocol/BUILD.bazel
@@ -1,5 +1,5 @@
-load("//dev:go_defs.bzl", "go_test")
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("//dev:go_defs.bzl", "go_test")
 
 go_library(
     name = "protocol",
@@ -7,8 +7,8 @@ go_library(
         "query.go",
         "searcher.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/searcher/protocol",
-    visibility = ["//visibility:public"],
+    importpath = "github.com/sourcegraph/sourcegraph/internal/searcher/protocol",
+    visibility = ["//:__subpackages__"],
     deps = [
         "//internal/api",
         "//internal/search/query",
@@ -19,7 +19,6 @@ go_library(
 
 go_test(
     name = "protocol_test",
-    timeout = "short",
     srcs = [
         "query_test.go",
         "searcher_test.go",
